feat(iOS): working app in all modes
