+++ lib/ios/AppDelegate/RNNAppDelegate.mm
@@ -21,58 +21,40 @@
 
 static NSString *const kRNConcurrentRoot = @"concurrentRoot";
 
-@interface RNNAppDelegate () <RCTTurboModuleManagerDelegate, RCTCxxBridgeDelegate> {
-    std::shared_ptr<const facebook::react::ReactNativeConfig> _reactNativeConfig;
-    facebook::react::ContextContainer::Shared _contextContainer;
-    facebook::react::RuntimeExecutor _runtimeExecutor;
-    std::shared_ptr<facebook::react::RuntimeScheduler> _runtimeScheduler;
-}
+@interface RNNAppDelegate () <RCTTurboModuleManagerDelegate, RCTCxxBridgeDelegate> {}
 @end
 
 #endif
 
 @implementation RNNAppDelegate
 
-#ifdef RCT_NEW_ARCH_ENABLED
-- (instancetype)init {
-    if (self = [super init]) {
-        _contextContainer = std::make_shared<facebook::react::ContextContainer const>();
-        _reactNativeConfig = std::make_shared<facebook::react::EmptyReactNativeConfig const>();
-        _contextContainer->insert("ReactNativeConfig", _reactNativeConfig);
-    }
-    return self;
-}
-#endif
-
 - (BOOL)application:(UIApplication *)application
     didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
     
-#ifdef RCT_NEW_ARCH_ENABLED
+    // Copied from RCTAppDelegate, it private inside it
     self.rootViewFactory = [self createRCTRootViewFactory];
     
+#ifdef RCT_NEW_ARCH_ENABLED
     RCTAppSetupPrepareApp(application, self.turboModuleEnabled);
+    RCTSetNewArchEnabled(TRUE);
     
     if (self.bridgelessEnabled) {
+        // Creating host instead of bridge
         self.rootViewFactory.reactHost = [self.rootViewFactory createReactHost:launchOptions];
         
         [ReactNativeNavigation bootstrapWithHost:self.rootViewFactory.reactHost];
         
         return YES;
     }
-#else
-    self.bridge = [[RCTBridge alloc] initWithDelegate:self launchOptions:launchOptions];
 #endif
+    // Force RN to init all necessary modules instead of copy-pasting code (ex. bridge)
+    [self.rootViewFactory viewWithModuleName:@""];
     
     [ReactNativeNavigation bootstrapWithBridge:self.bridge];
     
     return YES;
 }
 
-- (NSArray<id<RCTBridgeModule>> *)extraModulesForBridge:(RCTBridge *)bridge {
-    return [ReactNativeNavigation extraModulesForBridge:bridge];
-}
-
-#ifdef RCT_NEW_ARCH_ENABLED
 - (RCTRootViewFactory *)createRCTRootViewFactory
 {
   __weak __typeof(self) weakSelf = self;
@@ -87,9 +69,16 @@ static NSString *const kRNConcurrentRoot = @"concurrentRoot";
                                                    turboModuleEnabled:self.turboModuleEnabled
                                                     bridgelessEnabled:self.bridgelessEnabled];
 
+  if (!self.bridgelessEnabled) {
+      configuration.extraModulesForBridge = ^NSArray<id<RCTBridgeModule>> * _Nonnull(RCTBridge * _Nonnull bridge) {
+          return [ReactNativeNavigation extraModulesForBridge:bridge];
+      };
+  }
+    
   configuration.createRootViewWithBridge = ^UIView *(RCTBridge *bridge, NSString *moduleName, NSDictionary *initProps)
   {
-    return [weakSelf createRootViewWithBridge:bridge moduleName:moduleName initProps:initProps];
+    // Ignoring creation of default view, will initialize our custom UIViewController later
+    return nil;
   };
 
   configuration.createBridgeWithDelegate = ^RCTBridge *(id<RCTBridgeDelegate> delegate, NSDictionary *launchOptions)
@@ -97,7 +86,33 @@ static NSString *const kRNConcurrentRoot = @"concurrentRoot";
     return [weakSelf createBridgeWithDelegate:delegate launchOptions:launchOptions];
   };
 
+#ifndef RCT_NEW_ARCH_ENABLED
+    return [[RCTRootViewFactory alloc] initWithConfiguration:configuration];
+  }
+#else
   return [[RCTRootViewFactory alloc] initWithConfiguration:configuration andTurboModuleManagerDelegate:self];
+}
+
+#pragma mark RCTTurboModuleManagerDelegate
+
+- (Class)getModuleClassFromName:(const char *)name {
+    return RCTCoreModulesClassProvider(name);
+}
+
+- (std::shared_ptr<facebook::react::TurboModule>)
+    getTurboModule:(const std::string &)name
+         jsInvoker:(std::shared_ptr<facebook::react::CallInvoker>)jsInvoker {
+    return nullptr;
+}
+
+- (std::shared_ptr<facebook::react::TurboModule>)
+    getTurboModule:(const std::string &)name
+        initParams:(const facebook::react::ObjCTurboModule::InitParams &)params {
+    return nullptr;
+}
+
+- (id<RCTTurboModule>)getModuleInstanceFromClass:(Class)moduleClass {
+    return RCTAppSetupDefaultModuleFromClass(moduleClass);
 }
 #endif
