+++ lib/ios/Bridge/RNNBridgeEventEmitter.h
@@ -1,0 +1,4 @@
+#import "RNNEventEmitter.h"
+
+@interface RNNBridgeEventEmitter : RNNEventEmitter
+@end
