+++ lib/ios/Bridge/RNNBridgeEventEmitter.mm
@@ -1,0 +1,7 @@
+#import "RNNBridgeEventEmitter.h"
+
+@implementation RNNBridgeEventEmitter {}
+
+RCT_EXPORT_MODULE()
+
+@end
