+++ lib/ios/Events/RNNEventEmitter.h
@@ -1,20 +1,7 @@
 #import <Foundation/Foundation.h>
 #import <React/RCTEventEmitter.h>
 
-#ifdef RCT_NEW_ARCH_ENABLED
-#import <rnnavigation/rnnavigation.h>
-#endif
-
-@class RCTHost;
-
 @interface RNNEventEmitter : RCTEventEmitter
-#ifdef RCT_NEW_ARCH_ENABLED
-                            <NativeRNNEventEmitterSpec>
-
-- (void)setHost:(RCTHost *)host;
-
-@property(nonatomic, strong, readonly) RCTHost *host;
-#endif
 
 - (void)sendOnAppLaunched;
