+++ lib/ios/ViewControllers/Navigation/Delegates/StackControllerDelegate.h
@@ -1,4 +1,8 @@
+#ifdef RCT_NEW_ARCH_ENABLED
+#import "RNNTurboEventEmitter.h"
+#else
 #import "RNNEventEmitter.h"
+#endif
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
 
@@ -6,7 +10,11 @@
 
 @interface StackControllerDelegate : NSObject <UINavigationControllerDelegate>
 
+#ifdef RCT_NEW_ARCH_ENABLED
+- (instancetype)initWithEventEmitter:(RNNTurboEventEmitter *)eventEmitter;
+#else
 - (instancetype)initWithEventEmitter:(RNNEventEmitter *)eventEmitter;
+#endif
 
 - (BOOL)navigationController:(UINavigationController *)navigationController
                shouldPopItem:(BOOL)shouldPopItem;
